+++ src/treeView/repositories/repoItem.ts
@@ -1,7 +1,8 @@
 import os from 'os';
 import vscode, { ThemeColor } from 'vscode';
 import type { Dirtiness } from '../../commands/git/dirtiness/dirtiness';
-import type { Repository } from '../../store/repository';
+import type { Remote, Repository } from '../../store/repository';
+import { hasRepoRemoteWithUserAccess, isRepoOnDisk } from '../../store/repository';
 import type { TreeItemConstructor } from '../treeViewBase';
 import { TreeItem } from '../treeViewBase';
 
@@ -10,19 +11,24 @@ import { TreeItem } from '../treeViewBase';
 // TODO: Use GitHub icons (must resize them)
 // we may use repo-cloned as icon for template.
 function getIcon(repo: Repository): vscode.ThemeIcon | undefined {
-  if (repo.type === 'local') return; // No icon if local (sure?)
 
-  const args = ((): [name: string, color: string | undefined] => {
-    if (repo.isPrivate)
-      return ['lock', 'githubRepositoryManager.private'];
-    else if (repo.isFork)
-      return ['repo-forked', 'githubRepositoryManager.fork'];
-    else // is then public
-      return ['repo', 'githubRepositoryManager.public'];
+  const args = ((): { name: string; color?: string } => {
+    if (hasRepoRemoteWithUserAccess(repo)) {
+      if (!isRepoOnDisk(repo))
+        return { name: 'source-control', color: 'githubRepositoryManager.remote' };
+      if (repo.isPrivate)
+        return { name: 'lock', color: 'githubRepositoryManager.private' };
+      else if (repo.isFork)
+        return { name: 'repo-forked', color: 'githubRepositoryManager.fork' };
+      else // is then public
+        return { name: 'repo', color: 'githubRepositoryManager.public' };
+    }
+    return { name: 'symbol-folder', color: 'githubRepositoryManager.remote' };
   })();
+
   return new vscode.ThemeIcon(
-    args[0],
-    args[1] ? new ThemeColor(args[1]) : undefined,
+    args.name,
+    args.color ? new ThemeColor(args.color) : undefined,
   );
 }
 
@@ -32,40 +38,36 @@ const dirtyToMessage: Record<Dirtiness, string> = {
   dirty: 'This repository has local changes',
   error: 'An error has happened while getting dirtiness state! Read extension Output!',
   unknown: 'Checking if it\'s dirty...',
-
 };
+
 // + (repo.isTemplate ? ' | Template' : '') //TODO
 function getTooltip(repo: Repository) {
-  // TODO Maybe for windows it requires regex escape?
-  // os.homedir e.g. = linux: '/home/user'
-  const localPath = repo.localPath?.replace(RegExp(`^${os.homedir()}`), '~') ?? '';
 
   // the | &nbsp; | adds a little more spacing.
-  const R = repo.type === 'remote' ? repo : undefined;
+  const R = hasRepoRemoteWithUserAccess(repo) ? repo : undefined;
+  const D = isRepoOnDisk(repo) ? repo : undefined;
+  // TODO Maybe for windows it requires regex escape?
+  // os.homedir e.g. = linux: '/home/user'
+  const localPath = !D ? undefined : D.localPath.replace(RegExp(`^${os.homedir()}`), '~');
 
   const string = `
 |     |     |     |
 | --- | --- | --- |
 **Name** | &nbsp; | ${repo.name}`
 + (!R ? '' : `\r\n**Description** | &nbsp; | ${R.description ? R.description : 'No description'}`)
-+ `\r\n**Author** | &nbsp; | ${repo.ownerLogin}`
++ (!R ? '' : `\r\n**Author** | &nbsp; | ${R.ownerLogin}`)
 + (!R ? '' : `\r\n**Visibility** | &nbsp; | ${R.isPrivate ? 'Private' : 'Public'}`)
 + (!R ? '' : (R.languageName ? `\r\n**Language** | &nbsp; |${R.languageName}` : ''))
 + (!R ? '' : (R.isFork ? `\r\n**Fork of** | &nbsp; | ${R.parentRepoOwnerLogin} / ${R.parentRepoName}` : ''))
-+ (!R ? '' : `\r\n**Updated at** | &nbsp; | ${R.updatedAt.toLocaleString()}`)
-+ (!R ? '' : `\r\n**Created at** | &nbsp; | ${R.createdAt.toLocaleString()}`)
-+ (repo.localPath ? `\r\n**Local path** | &nbsp; | ${localPath}` : '')
++ (!R ? '' : `\r\n**Updated at** | &nbsp; | ${R.updatedAt}`)
++ (!R ? '' : `\r\n**Created at** | &nbsp; | ${R.createdAt}`)
++ (!D ? '' : `\r\n**Local path** | &nbsp; | ${localPath}`)
 // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-+ ((repo.dirty && repo.dirty !== 'clean') ? `\r\n**Dirty** | &nbsp; | ${dirtyToMessage[repo.dirty ?? 'clean']}` : '');
++ (!D ? '' : D.dirty !== 'clean' ? `\r\n**Dirty** | &nbsp; | ${dirtyToMessage[D.dirty]}` : '');
 
   return new vscode.MarkdownString(string);
 }
 
-
-type RepoItemConstructor = Omit<TreeItemConstructor, 'label'> & {
-  repo: Repository;
-  includeOwner?: boolean;
-};
 
 const dirtyToChar: Record<Dirtiness, string> = {
   clean: '',
@@ -74,16 +76,21 @@ const dirtyToChar: Record<Dirtiness, string> = {
   unknown: '?',
 };
 
-export class RepoItem extends TreeItem {
-  repo: Repository;
+type RepoItemConstructor<OnDisk extends boolean, R extends Remote> =
+  Omit<TreeItemConstructor, 'label'> &
+{
+  repo: Repository<OnDisk, R>;
+  includeOwner?: boolean;
+};
 
-  constructor({ repo, command, includeOwner, ...rest }: RepoItemConstructor) {
-    const repoName = includeOwner ? `${repo.ownerLogin} / ${repo.name}` : repo.name;
+export class RepoItem<OnDisk extends boolean = boolean, R extends Remote = Remote> extends TreeItem {
+  repo: Repository<OnDisk, R>;
+
+  constructor({ repo, command, includeOwner, ...rest }: RepoItemConstructor<OnDisk, R>) {
+    const repoName = repo.name;
 
     let description = '';
-    // if (Math.random() > 0.8) // Favorite
-    //   description += 'F ';
-    if (repo.dirty)
+    if (isRepoOnDisk(repo))
       description += dirtyToChar[repo.dirty];
 
     super({